Implemented infinite scrolling
fixed name of vars to camelCase and refactoring of gitHistory

Added a reference $this to $tableContainer and change the way to take the value of scrollHeight

Now the new rows are append with mustache

rewritten history panel code.
It doesn't load from git the list each time, the 3 parts of the panel script are now splitted in 3 functions and some little bugs are fixed.

gitHistory was remade with the suggestions of zaggino

Change the name of the function showCommitDialog to showHistoryCommitDialog

Fix an error related in issue #197

fix an issue with the style of the append commits in the table with the new commits

Fixed with @zaggino sugesttions

Author:    Jimmy Brian Anamaria Torres <captasfire@hotmail.com>