Added black and white avatars.
The problem with black and white avatars is that in a single project there could be thousand of people with a name starting with **M**, so them become just useless. In this case the colored avatars become very very very useful because are the only way (without internet) to have different avatars for each user.